requirements.txt
torch==2.3.0
opencv-python==4.9.0.80
streamlit==1.34.0
pillow==10.2.0
requests==2.31.0
numpy==1.26.4
pandas==2.2.1
torchvision==0.18.0
scikit-learn==1.4.2
matplotlib==3.8.3
seaborn==0.13.2
